Enforce no circular dependencies in CI with import-linter
Add acyclic_siblings contract for TraceLens, refactor pseudo-op extension
registry and graph capture merge to break import cycles, and run
lint-imports in the lint workflow.
