+++ .gitignore
@@ -41,6 +41,7 @@ venv.bak/
 .pytest_cache/
 .mypy_cache/
 .ruff_cache/
+.import_linter_cache/
 .hypothesis/
 .coverage
 .coverage.*
