+++ TraceLens/Trace2Tree/extensions/__init__.py
@@ -4,8 +4,8 @@
 # See LICENSE for license information.
 ###############################################################################
 
+from .pseudo_ops_registry import apply_pseudo_op_extensions
 from .pseudo_ops_utils import (
-    apply_pseudo_op_extensions,
     inject_pseudo_op,
     inject_pseudo_op_above_event,
     set_bookkeeping_attr,
